Refactor step 1 component for use in multiple places. Add registerAccountModal.
