+++ src/app/checkout/step-1/step-1.component.js
@@ -1,13 +1,6 @@
 import angular from 'angular';
-import 'angular-messages';
-import {Observable} from 'rxjs/Observable';
-import 'rxjs/add/observable/forkJoin';
-import find from 'lodash/find';
 
-import loadingComponent from 'common/components/loading/loading.component';
-
-import orderService from 'common/services/api/order.service';
-import geographiesService from 'common/services/api/geographies.service';
+import contactInfoComponent from 'common/components/contactInfo/contactInfo.component';
 
 import template from './step-1.tpl';
 
@@ -16,79 +9,24 @@ let componentName = 'checkoutStep1';
 class Step1Controller{
 
   /* @ngInject */
-  constructor($window, $log, orderService, geographiesService){
+  constructor($window){
     this.$window = $window;
-    this.$log = $log;
-    this.orderService = orderService;
-    this.geographiesService = geographiesService;
-
-    this.donorDetails = {
-      mailingAddress: {
-        country: 'US'
-      }
-    };
   }
 
-  $onInit(){
-    this.loadDonorDetails();
-    this.loadCountries();
-  }
-
-  loadDonorDetails(){
-    this.orderService.getDonorDetails()
-      .subscribe((data) => {
-        if(data['donor-type'] === ''){
-          data['donor-type'] = 'Household';
-        }
-        this.donorDetails = data;
-      });
-  }
-
-  loadCountries(){
-    this.geographiesService.getCountries()
-      .subscribe((data) => {
-        this.countries = data;
-        this.refreshRegions(this.donorDetails.mailingAddress.country);
-      });
-  }
-
-  refreshRegions(country){
-    country = find(this.countries, {name: country});
-    if(!country){ return; }
-
-    this.geographiesService.getRegions(country)
-      .subscribe((data) => {
-        this.regions = data;
-      });
-  }
-
-  submitDetails(){
-    if(!this.detailsForm.$valid){ this.$window.scrollTo(0, 0); return; }
-    let details = this.donorDetails;
-    this.submissionError = '';
-
-    var requests = [this.orderService.updateDonorDetails(details)];
-    if(details.email){
-      requests.push(this.orderService.addEmail(details.email));
+  onSubmit(success){
+    if(success){
+      this.changeStep({newStep: 'payment'});
+    }else{
+      this.submitted = false;
+      this.$window.scrollTo(0, 0);
     }
-    Observable.forkJoin(requests)
-      .subscribe(() => {
-        this.changeStep({newStep: 'payment'});
-      }, (error) => {
-        this.$log.warn('Error saving donor contact info', error);
-        this.submissionError = error.data;
-        this.$window.scrollTo(0, 0);
-      });
   }
 }
 
 export default angular
   .module(componentName, [
     template.name,
-    'ngMessages',
-    loadingComponent.name,
-    orderService.name,
-    geographiesService.name
+    contactInfoComponent.name
   ])
   .component(componentName, {
     controller: Step1Controller,
