+++ src/common/components/registerAccountModal/registerAccountModal.component.js
@@ -1,6 +1,10 @@
 import angular from 'angular';
 import 'angular-gettext';
+
+import contactInfoComponent from 'common/components/contactInfo/contactInfo.component';
+
 import sessionService from 'common/services/session/session.service';
+
 import template from './registerAccountModal.tpl';
 
 let componentName = 'registerAccountModal';
@@ -16,11 +20,20 @@ class RegisterAccountModalController {
   $onInit() {
     this.modalTitle = this.gettext( 'Your Contact Information' );
   }
+
+  onSubmit(success){
+    if(success){
+      this.onSuccess();
+    }else{
+      this.submitted = false;
+    }
+  }
 }
 
 export default angular
   .module( componentName, [
     'gettext',
+    contactInfoComponent.name,
     sessionService.name,
     template.name
   ] )
