+++ src/common/components/registerAccountModal/registerAccountModal.tpl.html
@@ -1,0 +1,15 @@
+<div class="modal-body">
+  <contact-info submitted="$ctrl.submitted" on-submit="$ctrl.onSubmit(success)"></contact-info>
+  <div class="row">
+    <div class="col-sm-5 hidden-xs">
+      <button class="btn btn-default" ng-click="$ctrl.onSuccess()" translate>Cancel</button>
+    </div>
+    <div class="col-sm-5 col-sm-offset-2">
+      <button type="button" class="btn btn-primary pull-right btn-block-mobile" ng-click="$ctrl.submitted = true" translate>Submit</button>
+      <button class="btn btn-link btn-block visible-xs" ng-click="$ctrl.onSuccess()">
+        <i class="fa fa-angle-left"></i>
+        <translate>Cancel</translate>
+      </button>
+    </div>
+  </div>
+</div>
